Add files via upload
This is the v2 version of the Spike Tools. It contains updates to the Gyro My Blocks. Specifically, it contains a new My Block called gyro turn to angle which will accurately turn your robot to the desired angle. The gyro drive to waypoint and gyro drive by distance have gyro turn to angle included. When they are called, they will first turn to the correct angle and then move in the correct direction. The advantage here is that now the distance traveled is accurately measured.