Add files via upload
Changed the gyro turn to angle to use a preset turn speed called gyro turn speed max. This will make it so the turns are done slow enough to move the robot but not fast enough to throw off the gyro. I have found that I can induce an error in the gyro if I turn too fast. This prevents that problem.

Note: gyro turn speed min was renamed to gyro turn speed max.